algunos cambios

+++ modelos/tecnico.py
@@ -1,10 +1,12 @@
 from dataBase import get_connection
 
+from .login import loggedAdmin, loggin
+loggedAdmin = True
 
-privilegiosAdmin = True
-
-def obtener_tecnicos():
-    conn = get_connection(privilegiosAdmin)
+def obtener_tecnicos(usuario, password):
+    if not loggin(usuario, password):
+        return ["Acceso denegado"]
+    conn = get_connection(loggedAdmin)
     cursor = conn.cursor(dictionary=True)
     try:
         cursor.execute("""
@@ -16,8 +18,10 @@ def obtener_tecnicos():
         cursor.close()
         conn.close()
 
-def agregar_tecnico(ci, nombre, apellido, telefono):
-    conn = get_connection()
+def agregar_tecnico(ci, nombre, apellido, telefono, usuario, password):
+    if not loggin(usuario, password):
+        return False
+    conn = get_connection(loggedAdmin)
     cursor = conn.cursor()
     try:
         sql = "INSERT INTO tecnicos (ci, nombre, apellido, telefono) VALUES (%s, %s, %s, %s)"
@@ -27,8 +31,10 @@ def agregar_tecnico(ci, nombre, apellido, telefono):
         cursor.close()
         conn.close()
 
-def eliminar_tecnico(ci):
-    conn = get_connection()
+def eliminar_tecnico(ci, usuario, password):
+    if not loggin(usuario, password):
+        return False
+    conn = get_connection(loggedAdmin)
     cursor = conn.cursor()
     try:
         sql = "DELETE FROM tecnicos WHERE ci = %s"
@@ -38,9 +44,11 @@ def eliminar_tecnico(ci):
         cursor.close()
         conn.close()
 
-def modificar_tecnico(ci, telefono):
+def modificar_tecnico(ci, telefono, usuario, password):
     # Modifica solo el teléfono del técnico normalmente no cambian de nombre o apellido
-    conn = get_connection()
+    if not loggin(usuario, password):
+        return False
+    conn = get_connection(loggedAdmin)
     cursor = conn.cursor()
     try:
         sql = """
